% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_mcp.R
\name{cv.cd.mcp}
\alias{cv.cd.mcp}
\title{cross validation for coordinate descent MCP}
\usage{
cv.cd.mcp(x, y, weights, type.measure = c("mse", "mae"), ..., nfolds = 10,
  foldid, trace.it = FALSE)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{type.measure}{loss to use for cross-validation. can be one of "mse" for mean squared error or "mae"
for mean absolute error}

\item{...}{Other arguments that can be passed to cd.mcp}

\item{nfolds}{number of folds for CV - default is 10. no smaller than 3}

\item{foldid}{an optional vector of values between 1 and nfold identifying whhat fold each observation is in. 
If supplied, nfold can be missing}

\item{trace.it}{If TRUE, then a printout that shows the progress is printed}
}
\examples{
set.seed(123)
n = 1000
p = 50
b = c(runif(10), rep(0, p - 10))
x = matrix(rnorm(n * p, sd = 3), n, p)
y = drop(x \%*\% b) + rnorm(n)

## fit lasso model with 100 tuning parameter values
res <- cv.cd.mcp(x, y, gamma = 4)


}

