% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_oglasso.R
\name{oglasso}
\alias{oglasso}
\title{Overlapping Group Lasso (oglasso)}
\usage{
oglasso(x, y, group, family = c("gaussian", "binomial"), nlambda = 100L,
  lambda = NULL, lambda.min.ratio = NULL, group.weights = NULL,
  standardize = TRUE, intercept = TRUE, dynamic.rho = TRUE,
  abs.tol = 1e-05, rel.tol = 1e-05, irls.tol = 1e-05, irls.maxit = 100L)
}
\arguments{
\item{x}{input matrix or SparseMatrix of dimension nobs \times nvars. Each row is an observation, 
each column corresponds to a covariate}

\item{y}{numeric response vector of length nobs}

\item{group}{A list of length equal to the number of groups containing vectors of integers 
indicating the variable IDs for each group. For example, group=list(c(1,2), c(2,3), c(3,4,5)) specifies
that Group 1 contains variables 1 and 2, Group 2 contains variables 2 and 3, and Group 3 contains 
variables 3, 4, and 5. Can also be a matrix of 0s and 1s with the number of columns equal to the 
number of groups and the number of rows equal to the number of variables. A value of 1 in row i and 
column j indicates that variable i is in group j and 0 indicates that variable i is not in group j.}

\item{family}{"gaussian" for least squares problems, "binomial" for binary response}

\item{nlambda}{The number of lambda values. Default is 100.}

\item{lambda}{A user-specified sequence of lambda values. Left unspecified, the a sequence of lambda values is 
automatically computed, ranging uniformly on the log scale over the relevant range of lambda values.}

\item{lambda.min.ratio}{Smallest value for lambda, as a fraction of lambda.max, the (data derived) entry
value (i.e. the smallest value for which all coefﬁcients are zero). The default
depends on the sample size nobs relative to the number of variables nvars. If
nobs > nvars, the default is 0.0001, close to zero. If nobs < nvars, the default
is 0.01. A very small value of lambda.min.ratio will lead to a saturated ﬁt in
the nobs < nvars case.}

\item{group.weights}{A vector of values representing multiplicative factors by which each group's penalty is to 
be multiplied. Often, this is a function (such as the square root) of the number of predictors in each group. 
The default is to use the square root of group size for the group selection methods.}

\item{irls.tol}{convergence tolerance for IRLS iterations. Only used if family != "gaussian". Default is 10^{-7}.}

\item{irls.maxit}{integer. Maximum number of IRLS iterations. Only used if family != "gaussian". Default is 100.}

\item{penalty}{Specification of penalty type. Choices include "gr.lasso", 
"gr.lasso.infty", "gr.mcp", and "gr.scad"}

\item{alpha}{alpha value for ridge penalty. Defaults to 1 for no ridge penalty.}

\item{gamma}{Tuning parameter of the MCP penalty; defaults to 3.}

\item{method}{Algorithm type for overlapping group lasso. Options include "admm" for the Alternating Direction
Method of Multipliers, "fista" for the Fast Iterative Shrinkage-Thresholding Algorithm. Defaults to "admm". 
"fista" is recommended when nvars >> nobs. More methods to come.}

\item{eps}{convergence tolerance for ADMM/FISTA iterations for the relative dual and primal residuals. 
Default is 10^{-4}, which is typically adequate.}

\item{inner.tol}{convergence tolerance for inner iterations. Does not apply for method = "admm". Default is 10^{-3}.}

\item{outer.maxit}{integer. Maximum number of outer (ADMM or FISTA) iterations. Default is 500.}

\item{inner.maxit}{integer. Maximum number of inner iterations. Does not apply for method = "admm". Default is 250.}
}
\value{
An object with S3 class "oglasso.fit"
}
\description{
Overlapping Group Lasso (oglasso)
}
\examples{
n.obs <- 1e5
n.vars <- 150

true.beta <- rnorm(n.vars)

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta

fit <- oglasso(x = x, y = y, group=list(c(1,2), c(2,3), c(3,4,5)))
}

